+++ Forge/src/main/java/com/christofmeg/brutalharvest/common/data/BrutalRecipeProvider.java
@@ -165,16 +165,6 @@ public class BrutalRecipeProvider extends BaseRecipeProvider {
                 .unlockedBy("tomato_seeds", has(ItemRegistry.TOMATO_SEEDS.get()))
                 .save(consumer, modLoc("tomato_seeds_from_tomato"));
 
-        ShapelessRecipeBuilder.shapeless(RecipeCategory.MISC, ItemRegistry.TOMATO_SEEDS.get())
-                .requires(ItemRegistry.ROTTEN_TOMATO.get())
-                .unlockedBy("tomato_seeds", has(ItemRegistry.TOMATO_SEEDS.get()))
-                .save(consumer, modLoc("tomato_seeds_from_rotten_tomato"));
-
-        ShapelessRecipeBuilder.shapeless(RecipeCategory.MISC, ItemRegistry.TOMATO_SEEDS.get())
-                .requires(ItemRegistry.TOMATO_SLICE.get())
-                .unlockedBy("tomato_seeds", has(ItemRegistry.TOMATO_SEEDS.get()))
-                .save(consumer, modLoc("tomato_seeds_from_tomato_slice"));
-
         ShapelessRecipeBuilder.shapeless(RecipeCategory.MISC, ItemRegistry.LETTUCE_SEEDS.get(), 2)
                 .requires(ItemRegistry.LETTUCE.get())
                 .unlockedBy("lettuce_seeds", has(ItemRegistry.LETTUCE_SEEDS.get()))